tools/pyproject.toml
[project]
name = "little-seeds-tools"
version = "0.1.0"
description = "GA4 / Search Console データ取得ツール"
requires-python = ">=3.10"
dependencies = [
    "google-analytics-data",
    "google-api-python-client",
    "google-auth",
    "python-dotenv",
]
